updated stageView
